BrunoTheBot.PDFAPI/API-PDF/requirements.txt
Flask==2.0.3
psutil==5.9.0
pdf2image==1.16.0
Pillow==9.0.1
pytesseract==0.3.8
werkzeug==2.0.3
